refactor: KakaoAuthenticationToken 별도 파일로 분리
